Minor changes to timestamp behavior.
